+++ types/search.ts
@@ -9,7 +9,7 @@ export interface PlayerSearchParams {
     name?: string;
     id64?: string;
     id3?: string;
-    tags: string[];
+    tags: Record<string, boolean>;
     after?: number;
     before?: number;
     sortBy: 'encounters' | 'lastSeen';